pyproject.toml
[tool.poetry]
name = "bumpbot"
version = "0.0.1"
description = "it bumps things"
authors = ["Luna Jackson <ojackson287@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
nextcord = "^2.0.0-alpha.10"
PyYAML = "^6.0"
loguru = "^0.6.0"
black = "^22.3.0"
isort = "^5.10.1"
flake8 = "^4.0.1"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
ipykernel = "^6.15.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
